Refactor SocketServer
